move Sidebar to layout folder
