+++ src/App.vue
@@ -32,7 +32,7 @@
 <script>
 import 'bootstrap/dist/css/bootstrap.css'
 
-import Sidebar from './components/Sidebar'
+import Sidebar from './components/layout/Sidebar'
 
 import AuthService from './services/AuthService'
