+++ src/components/layout/Sidebar/UserNav.vue
@@ -145,31 +145,31 @@ li.active > a::before {
 }
 
 .home.icon::before {
-  background-image: url('../../assets/home_icon-01.svg');
+  background-image: url('~src/assets/home_icon-01.svg');
 }
 
 .math.icon::before {
-  background-image: url('../../assets/math_icon-01.svg');
+  background-image: url('~src/assets/math_icon-01.svg');
 }
 
 .college.icon::before {
-  background-image: url('../../assets/college_icon-01.svg');
+  background-image: url('~src/assets/college_icon-01.svg');
 }
 
 .schedule.icon::before {
-  background-image: url('../../assets/calendar_icon-01.svg');
+  background-image: url('~src/assets/calendar_icon-01.svg');
 }
 
 .resources.icon::before {
-  background-image: url('../../assets/resources_icon-01.svg');
+  background-image: url('~src/assets/resources_icon-01.svg');
 }
 
 .training.icon::before {
-  background-image: url('../../assets/training_icon-01.svg');
+  background-image: url('~src/assets/training_icon-01.svg');
 }
 
 .profile.icon::before {
-  background-image: url('../../assets/profile_icon-01.svg');
+  background-image: url('~src/assets/profile_icon-01.svg');
 }
 
 .profile-info {
